2003.08.14_presto.zip

+++ presto/kernel/memory.h
@@ -9,20 +9,20 @@
 ////////////////////////////////////////////////////////////////////////////////
 
 typedef struct KERNEL_MEMORYPOOLSTATS_S {
-   short mempool_num_items;
-   short mempool_item_size;
-   short current_used_items;
+   unsigned short mempool_num_items;
+   unsigned short mempool_item_size;
+   unsigned short current_used_items;
    #ifdef FEATURE_MEMORY_STATISTICS
-      short current_total_bytes;
-      short max_used_items;
-      short max_requested_size;
+      unsigned short current_total_bytes;
+      unsigned short max_used_items;
+      unsigned short max_requested_size;
    #endif
 } KERNEL_MEMORYPOOLSTATS_T;
 
 ////////////////////////////////////////////////////////////////////////////////
 
 extern void kernel_memory_init(void);
-extern void memory_debug(int pool, KERNEL_MEMORYPOOLSTATS_T * stats);
+extern void memory_debug(unsigned short pool, KERNEL_MEMORYPOOLSTATS_T * stats);
 
 ////////////////////////////////////////////////////////////////////////////////
